refactor: extract TeamPageDataPreparer; wire into TeamService (Phase 1b)
